+++ CHANGELOG.md
@@ -23,6 +23,14 @@ BUG FIXES:
 - Fix error in `getMessageWithCode()` in CustomError when error cause is undefined.
 - Clean up data in cache DB when create request and create identity fails.
 
+## 0.5.4 (July 27, 2018)
+
+IMPROVEMENTS:
+
+- Separate body too large error into another error type.
+- Increase API body size limit to 3MB.
+- Increase message queue receiving size limit to 5MB.
+
 ## 0.5.3 (July 22, 2018)
 
 IMPROVEMENTS: